requirements/development.txt
asgiref==3.5.0
bandit==1.7.2
beautifulsoup4==4.10.0
behave==1.2.6
behave-django==1.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coverage==6.3.1
Django==4.0.2
django-prometheus==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
isort==5.10.1
lxml==4.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
parse==1.19.0
parse-type==0.6.0
pathspec==0.9.0
pbr==5.8.1
platformdirs==2.5.0
prometheus-client==0.13.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==3.5.0
tomli==2.0.1
unittest-xml-reporting==3.2.0
urllib3==1.26.8
uWSGI==2.0.20
